Moved the dice buttons to PlayboardScreen.kt and added the placeholder value '5'
