+++ app/src/main/java/at/aau/serg/websocketbrokerdemo/MainActivity.kt
@@ -76,7 +76,6 @@ class MainActivity : ComponentActivity() {
                             message = ""
                         }
                     },
-                    // onRollDice = { webSocketClient.sendMessage("Roll"); log += "Sent: Roll command\n" },
                     onLogout = {
                         auth.signOut()
                         val intent = Intent(context, AuthActivity::class.java)
@@ -103,6 +102,7 @@ class MainActivity : ComponentActivity() {
                 PlayboardScreen(
                     players = playerMoneyList,
                     currentPlayerId = userId ?: "",
+                    onRollDice = { webSocketClient.sendMessage("Roll")},
                     onBackToLobby = { navController.navigate("lobby") }
                 )
             }
