+++ app/src/main/java/at/aau/serg/websocketbrokerdemo/ui/PlayboardScreen.kt
@@ -1,5 +1,9 @@
 package at.aau.serg.websocketbrokerdemo.ui
 
+import androidx.compose.animation.animateColorAsState
+import androidx.compose.animation.core.FastOutSlowInEasing
+import androidx.compose.animation.core.animateFloatAsState
+import androidx.compose.animation.core.tween
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
@@ -7,20 +11,27 @@ import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.*
 import androidx.compose.runtime.*
+import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.rotate
+import androidx.compose.ui.draw.scale
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import at.aau.serg.websocketbrokerdemo.data.PlayerMoney
+import kotlinx.coroutines.delay
 
 @Composable
 fun PlayboardScreen(
     players: List<PlayerMoney>,
     currentPlayerId: String,
+    onRollDice: () -> Unit,
     onBackToLobby: () -> Unit
 ) {
+    var diceResult by remember { mutableStateOf("?") }
+    diceResult = parseDiceResult("Player 2 rolled 5")
     Box(
         modifier = Modifier
             .fillMaxSize()
@@ -40,13 +51,13 @@ fun PlayboardScreen(
         Column(
             modifier = Modifier
                 .align(Alignment.CenterStart)
+                .padding(start = 16.dp, top = 16.dp)
                 .fillMaxWidth(0.1f)
                 .fillMaxHeight()
                 .padding(16.dp),
             horizontalAlignment = Alignment.CenterHorizontally
         ) {
-
-            Text("🎲 Dice", fontSize = 24.sp, color = Color.Black)
+            DiceRollingButton("Roll Dice", Color(0xFF3FAF3F), onRollDice, diceResult)
         }
 
         // Player info column on the right (20% of screen width)
@@ -91,6 +102,16 @@ fun PlayboardScreen(
                 Text("Back to Lobby", fontSize = 18.sp)
             }
         }
+    }
+}
+
+fun parseDiceResult(newContent: String): String {
+    val diceRegex = "rolled (\\d+)".toRegex()
+    val matchResult = diceRegex.find(newContent)
+    return if (matchResult != null){
+        matchResult.groupValues[1]
+    } else {
+        "?" // Displays ? if no dice result was found
     }
 }
 
@@ -151,5 +172,61 @@ fun PlayerCard(
                 )
             }
         }
+    }
+}
+
+@Composable
+fun DiceRollingButton(text: String, color: Color, onClick: () -> Unit, diceResult: String) {
+    var isPressed by remember { mutableStateOf(false) }
+    var rotateAngle by remember { mutableFloatStateOf(0f) }
+
+    val rotation by animateFloatAsState(
+        targetValue = rotateAngle,
+        animationSpec = tween(durationMillis = 1000, easing = FastOutSlowInEasing)
+    )
+    val scale by animateFloatAsState(if (isPressed) 1.1f else 1f, animationSpec = tween(150))
+    val buttonColor by animateColorAsState(
+        targetValue = if (isPressed) color.copy(alpha = 0.7f) else color,
+        animationSpec = tween(durationMillis = 150)
+    )
+
+    Button(
+        onClick = {
+            isPressed = true
+            rotateAngle += 720f
+            onClick() // Hier wird diceResult im Log aktualisiert
+        },
+        modifier = Modifier.height(56.dp).scale(scale).rotate(rotation),
+        shape = RoundedCornerShape(8.dp),
+        colors = ButtonDefaults.buttonColors(containerColor = buttonColor)
+    ) {
+        Text(text, fontSize = 18.sp)
+    }
+
+    // Anzeige der geworfenen Zahl
+    DiceFace(diceResult)
+
+    LaunchedEffect(isPressed) {
+        if (isPressed) {
+            delay(1000)
+            isPressed = false
+        }
+    }
+}
+
+@Composable
+fun DiceFace(diceValue: String) {
+    Box(
+        modifier = Modifier.size(100.dp)
+            .background(Color.White, RoundedCornerShape(12.dp))
+            .padding(10.dp),
+        contentAlignment = Alignment.Center
+    ) {
+        Text(
+            text = diceValue,
+            fontSize = 36.sp,
+            fontWeight = FontWeight.Bold,
+            color = Color.Black
+        )
     }
 }
